// src/app/tool-bar/models/elements.ts
export class Elements {
    observaciones?: string;
    id_luminaria?: string;
    punto_luz?: string;
    vial?: string;
    numero?: number;
    lado_via?: string;
    distancia_eje?: string;
    s1?: number;
    s2?: number;
    centro_mando?: string;
    circuito?: 30;
    operativa?: string;
    altura?: 5.5;
    tipo_soporte?: string;
    marca_soporte?: string;
    modelo_soporte?: string;
    estado_soporte?: string;
    situacion_soporte?: number;
    tamano_brazo?: number;
    longitud_brazo?: number;
    orientacion_brazo?: number;
    tipo_luminaria?: string;
    marca_luminaria?: string;
    modelo_luminaria?: string;
    estado_luminaria?: string;
    tipo_lampara?: string;
    marca_lampara?: string;
    modelo_lampara?: string;
    estado_lampara?: string;
    cantidad_lamparas?: number;
    potencia?: number;
    equipo_auxiliar?: string;
    situacion_equipo_auxiliar?:string;
    orientacion?: number;
    alta?: boolean;
    usuario_alta?: string;
    fecha_alta?: string;
    modificado?: boolean;
    usuario_modificado?: number;
    fecha_modificado?: string;
    herramienta?: number;
    numero_registro?: 2283;
    envio?: string;
    fecha_envio?: string;
    id_centro_mando?: number;
    id_circuito?: number
  }